Introduce clean option for opera init CLI command
We are not introducing some hot stuff here. But don't feel bad
about this. Here we have a new functionality for the opera init.
I'll tell you a story. The developer that was well aware of the
powerful TOSCA standard once prepared his own CSAR for his
application. So then he wanted to deploy it of course. He saw the
simplicity of opera and decided to give it a try. Then he found
out that in order to deploy the CSAR he needs to initialize it
first. He initializes it and quickly remembers that he has forgotten
to add some changes to the files in the CSAR. He applies the changes
and initializes again but the CSAR got corrupted as the old and new
CSARs got mixed and cannot be deployed properly. The developer said to
himself: "Oh, if I could have an option that would clean the storage
for me before the CSAR initialization or that it would warn me if the
CSAR was already initialized so that I can do no harm.". Well the
story will have a happy ending because we added opera init --clean/-c
flag that solves that problem, which was originally spotted in #105.