Add a CI routine to test the examples
Where to begin? Tests, examples, more examples in the other repository.
Users can experience a total confusion. But we will not allow this to
happen with opera. That's why we first opened #93 issue as we wanted to
clean up the mess and move all the existing opera examples to the
integration test folder (including the examples in this repository and
also the examples from xopera-examples repository). However, this would
not be right. It would be better to keep the schema as it was designed
for new and experienced users. The /examples folder contains mostly
simple TOSCA templates that can be tested right away and can prove that
the tool actually works. For more complex tests we still have a
separate repository for now. But this is not the goal on this commit.

With these changes we are adding CI/CD tests for opera examples. Opera
is constantly being developed, so we soon realized that the examples
must be kept up to date with the newest features and bugfixes. This is
a lot easier if we establish a new CI routine to test the examples
every time we apply changes and push something new. The bash script
(examples/runme.sh) that is used by the CI can be also tested manually.
We excluded ento-to-end examples (like OpenStack and nginx) from testing
because they require a special set of prerequisite resources.