Allow compressed CSAR deployment with opera deploy
These changes correspond to the debate within the #107 issue. There we
have agreed that opera init command is unneeded at the moment as it
does not do anything different from opera deploy (apart from just
skipping the last deployment step). We can achieve the same behavior
with opera deploy by adding the compressed CSAR initialization logic.
And that's exactly what we did. Users are now able to deploy TOSCA
service templates, uncompressed CSARs or compressed CSARs with opera
deploy. We thought about the possibility that some users would still
initialize the service template/CSAR first and the deploy, so we kept
the option to run just opera deploy after the initialization.

With all this, we have know shortened the path for the compressed CSAR
deployment as users can now achieve the same thing with only one CLI
command instead of running two. This means that when deploying zipped
CSARs, opera init and opera deploy are not needed to go together on
this journey. Opera deploy has become self-sufficient and an alone
traveller. And he travels the fastest who travels alone, right?