Equip opera CLI with CSAR unpackaging command
The issue #124 presented the importance of having both, packaging and
unpackaging commands supported within the opera TOSCA orchestrator.
So be it. In this commit we are boosting up opera's CLI commands
by introducing the new opera unpackage command, which will serve for
unpacking (i.e. validating and extracting) the compressed TOSCA CSAR
files. The opera unpackage command receives a compressed CSAR as a
positional argument. It then validates and extracts the CSAR to a
given location. The available command switches are:

  1. --destination/-d is the folder location where the CSAR will be
     extracted to. If the specified path does not exists, it will be
     created during the unpackaging process. The flag can be omitted
     and this means that the output dir will be named using UUID.

  2. --format/-f is the option that specifies the compressed file
     format that will be used to extract the CSAR. Currently, there
     are two choices here: 'zip' or 'tar', where 'zip' is the
     default compression method.