Allow passing inputs in JSON format
Inputs can be very important for the orchestration process as they
often include various secrets and variables that are used to
authenticate and to deploy applications. Up until now inputs were
passed to the orchestrator as a YAML file with key-value pairs.
To make our TOSCA orchestrator more flexible we can extend inputs
format (this was originally explained in the issue #122). So, with
these changes we are adding the possibility to supply inputs also
in JSON format. This was not very hard because opera already stores
inputs as a JSON file in its storage folder (.opera by default) and
therefore we only had to modify the things for certain CLI commands.
This is the part where --inputs-format CLI option has been brought to
life. It can be used to specify the desired YAML or JSON inputs format
(YAML is used by default) for validate, init or deploy CLI command.